pxe: add vendor option (43) suboption (70) for machine UUID
DHCP can provide the machine's UUID to the booting client for use as an
URL parameter to the PXE generator so it can match the correct machine
by UUID instead of doing hacky reverse IP lookups or some other thing.

It may be useful in the future to provide a kind of "deploy key" via
DHCP and signed by the hosts TPM public key so that we only boot a host
we think we're booting.