dhcp: start handling vendor_class to generate the correct location
This updates the DHCP component to parse DHCP option 60 for the Vendor
Class in RFC 2132 section 9.13.  We need to extend the client detection
code to include DHCP option 93 in RFC 4578, section 2.1 as it appears
clients will send both of these, and I'm not sure what to do yet if they
disagree.