pxe: METAL-24: add ephemeral image builder
This adds a subtree for `mkosi` that is used by `cargo make ephemeral`
to build an ephemeral image.  The build process needs to be run as root
so right now it's excluded from the CI/CD until we can test it.  The
resulting build will produce a Unified Kernel[1], as well as a Root
Filesystem.  This will be booted with iPXE to run the ephemeral image.

For now, this means we can test the image in systemd-nspawn to develop
some more of the carbide featureset without booting an entire machine.

TODO:
- Functionality to add carbide build artifacts to the image
- Turn on the signatures and dm-verity stuff to make this image be
  secure-boot'able and have ipxe validate signatures
- Add build id to the ephemeral image based on CARGO version
- Make the image run cloud-init to drive the carbide system