interfaces: properly type behavior on absent subnet
Rather than use a bool, add an enum that defines what to do if we're
selecting an address on a subnet without that address family.  Ignore if
we just return "no address" or fail if we want an actual error (i.e. we
want an IPv4 address on the network with no IPv4 address).