Add bits to ensure EFI PXE HTTP boot works
This commit adds the proper support for PXE booting via HTTP. This
requires a specific string to be sent back in the vendor options,
which is currently hard coded. It also fixes up the readme to ensure
the proper things are built prior to trying to PXE boot the qemu
instance.

Closes: FORGE-228