Put the interface uuid into the dhcp response
This commit adds the interface id from the DB to both the view and the
dhcp response. It also fixes the uuid brokenness in dhcp by using this
uuid instead of the machine id. It also adds configuration for the
HTTPClient as a sneaky little side effect of the commit.

Relates: FORGE-234
Closes: FORGE-240