Modify mkosi build to pull in cli tools
This commit adds the proper systemd files and wrapper script that will
ensure the carbide-cli is run after networking starts in the
image. The wrapper script uses parameters from /proc/cmdline which is
supplied by our pxe server. These values are used for the URI of the
phone home server as well as a uuid associated w/ the server.

Relates: FORGE-103