Store machine topology on discovery
This commit is the server side impl of a machine topology. It is
stored in the db as jsonb. It also associates the nic's uuid to the
machine. This does not implement any additional logic to get the
topology back. This is only the initial save of the data during
discovery.

Closes: FORGE-107
Closes: FORGE-105