Change boot instructions based on machine state
This commit does a few things to enable state based booting of
machines. It first re-adds an optional machine back to the ipxe
service. It fleshes out the getter for machines in the api as
well. And finally it does a little string matching on the machines
state such that it can make a determination on what instructions to
give to ipxe.

Closes: FORGE-324