feat: Mark DPU's no longer reporting network health as unhealthy
If `forge-dpu-agent` hasn't reported a network health observation in
over 5 minutes (~10 missed calls) we mark the DPU's network as
unhealthy. Later that will prevent instances being allocated to this
machine. As soon a the next healthy report arrives we go back to
healthy.

If there is already an instance on that machine we still mark it as
unhealthy, but nothing happens to the instance - apart from presumably
it has networking problems. Later Cloud UI might surface the information
for more rapid problem solving.

This matches how we report Prometheus metrics. After 5 minutes of missed
health reports we mark the DPU as down in it's Prometheus metric.

Includes test.