test: Allow to create multiple ManagedHosts in unit-tests
This change allows to create multiple managed hosts (and thereby also multiple instances)
in a single unit-test.
Each DPU and Host will get a unique set of hardware properties (mac addresses and serial
numbers assigned), which are generated from a pool. The previously pre-assigned mac
addresses and ip addresses are no longer used.

In addition to that the `create_host` and `create_managed_host` functions in test will
now also correctly simulate the DHCP request of the hosts BMC, which is actually the
first step in the discovery process. The IP that will get assigned by carbide will then
be submitted as part of the machines BMC hardware details, instead of a previously
used arbitrary API that mismatched the actual interface.

The added test `test_discover_2_managed_hosts` validates that we can actually discover
2 distinct hosts. It failed with the following error before the change:
```
// 2023-09-18T23:48:54.108841Z ERROR carbide: Database Error: error returned from database: duplicate key value violates unique constraint "machines_pkey" file=api/src/db/machine.rs line=418 query=INSERT INTO machines(id, controller_state_version, controller_state, network_config_version, network_config) VALUES($1, $2, $3, $4, $5) RETURNING id. location=./src/api.rs:1809:17 handler='2: <carbide::api::Api<C1,C2> as rpc::protos::forge::forge_server::Forge>::discover_machine::{{closure}}'
// thread 'test_discover_2_managed_hosts' panicked at 'called `Result::unwrap()` on an `Err` value: Status { code: Internal, message: "Database Error: error returned from database: duplicate key value violates unique constraint \"machines_pkey\" file=api/src/db/machine.rs line=418 query=INSERT INTO machines(id, controller_state_version, controller_state, network_config_version, network_config) VALUES($1, $2, $3, $4, $5) RETURNING id.", source: None }', api/tests/common/api_fixtures/dpu.rs:275:10
```