feat: Remove tenant neighbors health check
Remove tenant neighbors being added to the base configuration and health
checked as part of base FRR configs.

The tenant neighbors might not be used, we have no way of knowing, so
they might be marked as failed. We now allow that.

Does not apply to admin network.

Thanks Michael Basnight and Nick Vetro for the collab.