feat(agent): Fix how agent restarts
Previously we tried to use `systemctl restart` in a package `postinst`
action. forge-dpu-agent calls the `apt-get install` which calls `dpkg`
which calls the `postinst` script. So when postinst called `systemctl
restart` it killed a parent process. That meant the install did not
complete. Our package was left in "half-installed" state and `dpkg`
complains on subsequent runs.

Now we revert to initial design. `forge-dpu-agent` exits once the new
version is installed. `dpkg` does not attempt to restart us. `systemd`
will restart us because our unit has `Restart=always`.

Adjust the `postinst` script to only run when it's meant to, and make it
more of a real postinst script.

Commit also adds handling of TERM signal to forge-dpu-agent. That is
what `systemctl restart` (or systemctl stop, and unix in general) uses to stop
the process. If we don't handle it we get default handling which stops
our process immediately, potentialy in the middle of something important.
Now we only stop in the quiet period between main loop iterations.