feat: Allow to configure instances to always use custom iPXE script
The current behavior of Forge is to supply the tenants custom iPXE script
only on the first boot attempt. In the following boot attempts Forge will
only provide an "exit" script - with the goal to persist previously installed
OS and not trigger reprovisioning of the instance.

For tenants which manage the decision whether to reprovision or not
via their own control plane this behavior is not sufficient. Their instance
should always PXE boot, and its up to the tenant to decide the next step.

This merge request introduces the flag `always_boot_with_custom_ipxe`
which allows for that. If the flag is specified during instance creation,
the each boot attempt will lead to the custom iPXE script being applied.

Fixes https://jirasw.nvidia.com/browse/FORGE-2324