feat: Return pf_guid for infiniband interfaces in addition to guid
If a tenant configures an instance for infiniband, they configure it by device name.
Forge then selects a hardware device based on the device name, which maps
to one particular device path on Linux.

The tenant needs to be aware of that path in order to apply configurations
(e.g. configure VF devices).
The mapping algorithm that Forge uses to map from device names to hardware
devices is deterministic, therefore the tenant could perform the mapping from
device name and device instance to hardware GUID themselves. To make tenants
life easier, we however discussed that we could also return the GUID of the
associated PF device. In that case the tenant can look up the device by searching
for this GUID.

This change implements this improvement by returning `pf_guid` in the Infiniband
interface status.

In addition it reorganizes Infiniband related unit-tests:
- All tests related to instance creation while using Infiniband have been moved
  to `ib_instance.rs`.
- Tests are using the public APIs in order to create instances and retrieve their
  status.