feat(agent): Log a message once we are stable
Currently we don't have any noisy "all is well" logs, which is good, but
it means the last log you usually see is the last problem we had. Now
log a single line the very first time HBN reports as healthy and we have
applied the network configuration and it hasn't changed since last
iteration (aka stable state). This will give us a timestamp of when
things started being OK.

Also re-generate the forge-admin-cli autocomplete for
https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/1371