fix: Always show that an Instance is in Terminating state after ReleaseInstance has been called
If `ReleaseInstance` is called on an instance that did not even finish
provisioning, the FindInstance API still showed an instance state of `Provisioning`.
That did not provide users any confirm on whether the deletion was actually scheduled.

To fix this problem, we always overwrite the instance state to `Terminating` after a
Release call had been received.
We already use the same pattern for Network Segments and IB Partitions which will
immediatly switch state to Terminating after any deletion call - even if the state
machine first has to process additional creation steps.