chore: Move StateHandler metrics into StateHandler context
The StateHandler `ctx` parameter will soon get extended to hold additional data
per iteration which needs to be generic (specific to one specific type of state controller).
E.g. it would be able to preload data that is used for acting on all machines.

Since the Context will become generic anyway, we can now als move
the metrics that are produced as the result of a state handler iteration
there. This removes one parameter in `handle_object_state`.