chore: update scout/qcow images
mkosi has had a ton of changes in the last 18 months, so we're updating
mkosi and reworking the postinstall script and the qcow imager to be
more standardized into how mkosi works.  This also upgrades the scout
and mkosi image from Debian 11 to Debian 12.  This is required for
BlueField-3 support.  The older kernel doesn't have the correct kernel
images to detect the card.

- Separate mkosi into `mkosi.profiles` (scout & qcow-imager)
- Update cargo build to use cargo to detect the directory instead of
  heuristics about whether we're in CI or not.
- Convert the postinst script for cuda-drivers and the dell tools to be
  native files copied into the mkosi filesystem
- Build images as UKI images instead of separated kernel/initrd so that
  we don't need to have dracut livenet support.
- Move the group creation into /etc/sysusers.d
- Move the service enable/disabled into /etc/systemd/systemd-preset
- Update the API iPXE instructions to boot the new scout UKI.
- mkosi doesn't seem to require sudo anymore 🎉
- Delete all the makefile code for different Root CAs

ci-run-complete-pipeline

Signed-off-by: Andrew Forgue <aforgue@nvidia.com>