perf: Don't load unnecessary DHCP entries when fetching MachineInterfaces
The existing code was loading all DHCP entries from the DB, even if just a single
MachineInterface was requested.
This change will only load entries for the Machine interfaces that are affected.

It also adds an optimization where we forward `All` queries for
MachineInterfaceAddresses and DhcpEntries, if `All` MachineInterfaces had been
queried. Only fi a specific set of interfaces is queried, the IDs inside the set
are forwarded to sub-queries. This avoids transferring big lists if not necessarily
required.