feat: Add SiteExplorer subsystem
The SiteExplorer component within carbide-api will periodically explores
all endpoints which retrieve their IP address from Carbide by DHCP.

It tries to explore the endpoints using redfish APIs and using well-known
credentials and persists the discovered information in the database.

The format of the discovered information follows the redfish standard
and information hierarchy (Systems, Managers, etc).
The data is however flattened.

The explorer runs periodically (default: every 5min), and explores up to
a configurable amount of endpoints (default: 10) per run. Newly discovered
endpoints are explored first. Otherwise the oldest exploration results are refreshed.
This mechanism limits the load that the Site Explorer introduces.

Site Explorer can be enabled via config file, and is disabled by default.

The Site Exploration Report is persisted in the database, but not yet
available via APIs (will be a follow-up).

Site Explorer emits metrics about the amount of successfully explored
endpoints and endpoints which could not be explored.