feat(dhcp): Put the MTU API gave us in the DHCP Offer
We already had the network segment MTU in grpc response from API server.
Now we include that in the Offer as Option 26 "interface-mtu".

With this change we are no longer always sending 1500 as the MTU. It
will depend on how the network segment was defined in the config
file (admin and underlay) or in create_network_segment (tenant).