feat: Generate MachineId for DPUs in SiteExplorer report
If enough information is available to generate a MachineId for an explored endpoint
in SiteExplorer, the persisted report will generate the MachineId.

The functionality is for now only implemented for DPUs, since Hosts require
TPM data.
It will also require a new version of DPU firmware which actually exposes the
serial number via redfish.