chore: Upgrade all opentelemetry libraries to 0.21
A former change already upgraded parts of the project to OTEL 0.20, but didn't
upgrade all other parts.
This lead to duplicate compilation of OTEL and dependencies.
This change upgrades everything to the same version.

A small change for custom metric views was required, since OTEL doesn't
really support regexes for them - just globs. Since 0.20 also supported
regexes, this had to be changes.