fix: bump the host wait time for dgx
The DGX's are slow to boot and have lots of interfaces to cycle
through. Until we have spare time to clean up how the DGX is set up
prior to prov, we need to bump this timeout as it is currently riding
a line that some DGX successfully reboot in the original window and
some do not.