feat(api): (re)-add tokio-console
Marked as Draft because I don't want to merge this until after the break when I can monitor it in an real environment.

---
Uses my fork of the tokiko-console project.

The changes from upstream are:
- Reduce default retention to 6 seconds - so we don't need the one
  minute retention we were setting previously
- Don't retain poll_ops forever. For obvious reasons.
- If we have too much old data to fit in a message, keep less old data.

Links to the changes from upstream here:
https://github.com/grahamking/console/blob/main/README.md

My hope is that these will be merged and we can switch back to using
the official version soon.

Reverts commit 9075087cde4eefceb8e5109c8491ea4e3c43266f to restore tokio-console.