feat(agent): Health check for forge_admin user, and create it
New health check: ipmi user `forge_admin` must exist.

If that health check fails we report the failure, then we try to create it.

Creating the ipmi user uses code that used to be in `scout` and is now shared.
If creation succeeds forge-dpu-agent will notice on it's next run and succeed the
health check.