chore: fixing some small agent spelling/grammar so i can have my first MR
To make sure I understand (and can experience) the MR workflow, and that all
of my stuff works, I'm making a reeeeeeally basic MR. All I did was look
through the agent codebase to find small instances of spelling and grammar
typos with comment blocks and lines. Nick is definitely going to be like smh.