feat: Emit metrics when machines/partitions/subnets stay in a state for longer than desired
This change adds another metric which should make it easier to identify stuck instances.
Besides the `forge_machines_per_state` metric, we will emit
a `forge_machines_per_state_above_sla` metric which shows the exact amount of
machines per state that have been stuck in a state for longer than anticipated.

The maximum amount of time an object can stay in a certain state is encoded in the function
`state_sla` for each state controller. The definitions in this MR are:
- Objects in `ready` state can stay there indefinitely
- Creation and Termination for Subnets and IB partitions should happen within 15min
- All machine state changes should happen within 30min

The new metric should therefore be easier to use use and understand in Grafana than the
current metric for identfiying stuck objects:
```
0.2 * 0.5 * (sum by (state, substate)(increase(forge_machines_time_in_state_seconds_count{forge_site="${ClusterName}", state!="ready", substate!="ready"}[5m])) - sum by (state, substate)(increase(forge_machines_time_in_state_seconds_bucket{forge_site="${ClusterName}", le="2500", state!="ready", substate!="ready"}[5m]))) # *0.2 since we look at 5 minute buckets, *0.5 since the state handler runs up to twice per minute
```

The existing metric for stuck objects also used an SLA of 41min for each possible state transition,
while the new mechanism allows to configure a limit per state. This allows to have a shorter SLA
for state transitions which are expected to happen immediatly (e.g. subnet becoming ready),
and a longer SLA for transitions which are expected to take some time (e.g. locking down a Machine).

A gap in the current implementation of this mechanism is that objects which re-enter
the same state (due to retries) will not be identified as being "above SLA", since
re-entering the state will reset the timestamp. This will be addressed in future work.