feat: SiteExplorer - Identify and store Host PF MAC
So far SiteExplorer linked the Host and DPU. However it was hard to identify
which MAC address/interface of the Host is actually provided by the DPU.

With this change, the Hosts PF interface provided by the DPU will be
identified (by MAC address prefix matching) and stored in the database.

The information is also accessible via forge-admin-cli and web UI.
This makes it easier for users to search for a ManagedHost pair that
uses a certain MAC address.