chore: update pxe to use agent_config to back user-data
When I was poking around code, I noticed the pxe server was just building
up a string for the agent config toml that it passes back, and it *seemed*
like a nice opportunity to instead back it with an `AgentConfig`.

Three questions I do have are:
- Are making `hostname` and `mac_address` public ok here?
- Would it be better if there was an `AgentConfig::FromPxe` constructor?
- Should I pass through the client cert/key details from `RuntimeConfig`?

This also might be overkill, since the string is pretty basic, but I
do think that if we already have a struct backing the data, and a way to
serialize the data to TOML, it would be nice to leverage it!