fix: Emit correct forge_machines_per_state_above_sla metric for Machines in retried state
The existing logic for determining whether a machine was longer in a state than allowed
only looked at how long the machine was exactly in the current state.
However some Machine states like BootingWithDiscoveryImage are retried, where a retry
causes the state to be overridden and the timstamp to be reset. Therefore we never had
been in a single `BootingWithDiscoveryImage` state for more than 30min, and the SLA
logic did not show machines as stuck.

The new logic fixes that by lettting the function whether a Machine is stuck examine
the actual state in addition to the state timestamp.