chore: move [lib] crates into a common directory
I kicked off some discussion based on this MR (https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/1687), which led to
some more discussion in the team channel about code organization, where
Andrew said:

```
if you want to do the work to move things into a crates dir i would support it
```

Graham had mentioned the dir being called `common`, so that's what I'm doing.

Thanks to Bill for clarifying a few things for me -- originally I was going
to make a single `common` crate with a bunch of modules, which would have,
in turn, been a bunch of small code changes to prefix `common::` all over
the codebase.

Instead, we've just got some TOML file changes.

A few things to note:
- I made sure to collect all of the "local dependencies" together, since
that seemed to be something we were trying to do.
- The "local dependencies" section had varying comments, so I'm attempting
to standardize on a `# [local-dependencies]` comment -- easier to grep for.
- I had to update some paths in `kea.rs` for building + loading `libdhcp.so`,
which yelled at me during my first `cargo test`.