test(agent): Wait for mock grpc server to start
Previously we waited exactly 500ms for the mock grpc server to be ready
to respond on it's TCP listener. This was occasionally too short for CI.
It was always too long for my local dev.

Now we poll an `/up` endpoint every 100ms for up to 2 seconds.

Introducing the `reqwest` library just for this was admittedly a bit
lazy but it's dev-only, and it's already in Cargo.lock. Happy to replace
it with hyper-only if requested.