chore: Cleanup old Forge Native Networking
Six months ago we put in templates for FNN that never ended up getting
used. FNN turned out to be a bit more complicated than that.

Now that we're doing it again, with NVUE, cleanup the old version to
avoid confusion.

The only part I am not cleaning up is the Postgres enum
`network_virtualization_type_t`, because delete values is not supported
there.