chore: add tcpdump and nc to the discovery image
We are/were having an intermittent issue in `az02` where gRPC calls to
`carbide-api` (register, query_api/control, etc) are getting a connection
reset.

I jumped into the virtual console of an unassigned `Ready` machine to
troubleshoot, but neither `nc` or `tcpdump` were installed. I figured
I could then jump onto a DPU and run a `tcpdump` there, but since all
traffic is offloaded by hardware (except for a few things that get
punted to the CPU), I couldn't actually `tcpdump` there either; the
answer was to jump on the switch.

It would be nice if we had these tools in the discovery image if
we need to do any sort of provisioning troubleshooting related to
connectivity.

Also: I think this is the right place to put this, and, once/if it
goes in, I have no clue how to then do a build of a new image and
make it available to machines the next time they are released.