feat(agent): Enable core dumps in systemd file
Part of /proc/sys/kernel/core_pattern

The agent hung so we wanted to get a stack trace of where it's stuck, we
tried sending it the usual `ABRT` signal, but that was disabled. So
enable it for next time.

Note that I removed the `+` at the start of `ExecStart`. The plus put us
in "privileged mode" where none of the other restrictions in the systemd
file would apply. We don't have any other restrictions in this file,  so
removing it should be fine. Core dumps don't work on processes with
that `+` set, presumably for security reasons.

On startup delete all but the most recent core file, to prevent
unbounded growth.