feat(agent): Have systemd restart the agent if it hangs for too long
- Switch our service type to `notify`, which means we tell systemd when we
have started (and optionallly when we are stopping).

- Use the [systemd
watchdog](https://0pointer.de/blog/projects/watchdog.html). We ping it
regularly. If we don't ping in `WatchdogSec` systemd uses SIGABRT to
capture a core dump then restarts us. The "ping" is writing a well-known
string to a SOCK_DGRAM that systemd gives us.

Restart period is 10 minutes because:
- That's longer than carbide-api's HeartbeatTimeout so we'll still get
  metrics that something is wrong
- Michael requested a bit of time to get onto the machine and confirm.