fix: Apply timeout on state handler iterations
The time that the state handler had so far for inspecting and advancing the
state of each object was unbounded. However the state handler needed to
wait for state handling for all objects to finish before the next iteration could
start. Therefore any slow object would lead the state handler to be come
slow - or would fully block it.

This behavior becomes especially problematic since our postgres library
sqlx does not use any timeouts for queries. Thereby a blackholed TCP
connection between client and postgres server could make the state
handler block indefinitely.

To avoid this deadlock, this change applies an overall timeout to each state
handler run. This works since each handler is a "future". It would however
lead to problematic behavior if the state handler uses `spawn_blocking`,
since the blocking call would run on a background thread and could not
be cancelled.

A better solution than this change would be to have actual query timeouts
in sqlx. That would however require a fix in the dependency.