chore: wire up BmcInfo with the optional port as exposed by the redfish lib
The background for this is that I went to write unit tests for https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/1722,
and when I did, I then ran into an issue where the `"ip"` field of `BmcInfo`
had `"localhost:1266"` as the value.

The reason is so unit tests can hit the local `bmc-mock` server, but the
problem is this breaks the ability to generate a DNS view based on IP
addresses; the addition of a `:<port>` breaks the `->inet` conversion.

Soooo.. just embrace the fact the `redfish` client library has first-class
support for a `port` (`Option<u16>`), add an optional `port` to `BmcInfo`,
and wire it on through!

This also adjusts the `"localhost:1266"` (et al) unit test data to leverage
this new integration.

Once this goes in (assuming this all seems reasonable), I can then finish
the unit test work in the aforementioned MR.