chore: consistenty borrow the admin cli Config throughout
When I was tweaking the `handle_show` commands earlier today, I noticed
there was some inconsistency throughout the `admin/` codebase around
how `Config` (generally named `api_config`) was being passed around:
- There were cases of it being moved, then moved again.
- There were cases of it being moved, then passed as a reference.
- There were cases of it being passed as a reference, and used as a reference.
- There were cases of it being passed as a reference, and then `clone()`d.

No judgement fwiw, I'm just excited there might be something I can do in here,
and undoubtedly in a number of other places.

All said, I'm taking the approach of ownership staying where `Config` is
first initialized (in `main.rs`), and then subsequently passing by reference
throughout.

Some reasoning behind this is:
- I generally am one to not transfer ownership unless I have to.
- Seeing variables passed by reference makes it easier for me to comprehend
that it's an existing value, instead of [possibly] a new one.
- It reduces the need for potential `clone()` calls later.
- It's consistent!

We could probably discuss the performance tradeoffs of doing a move vs.
passing a reference, but I'm not really here for that; this is mainly
just about consistency! ..and, for me, it helps with readability and
maintainability.