fix: Avoid repeated compilation of version crate and all dependents
When the version crate was moved under `common`, the path relative to `.git/HEAD`
wasn't adjusted. As a consequence, each `cargo build` invocation rebuilt the version
crate and everything that depended on it - even if nothing in the code changes.

This change fixes the path and adds an assertion that avoids that those path go out
of sync in an unnoticed fashion.