fix: Don't delete dpu agent `upgrade` test when running it
The test binary deleted itself while running the upgrade command. This change detects
that the upgrade command is overridden and in that case prevents moving the binary.

It also moves the definition of the override command into a different section of the config
since the setting is not "Machine" related. This is not a breaking change since real DPUs
don't specify the setting and a default for the setting is provided.