fix: Fix SiteExplorer crashes when trying to extract the DPU base MAC
The new code to extract the base mac used various unwrap statements which lead the siteexplorer task to panic in case the information for the DPU wasn't retrieved or was invalid.

This change adds checks to every optional piece of information, as well as enforces that the MAC is a valid MAC.

Unit-tests have been added.