feat(api): Config setting to enable/disable NVUE
Previously we used a database entry per-VPC for intance networks, and
hard coded service (admin) networks.

Now it's a single carbide-api flag for both. If we enable it and the
DPU doesn't support NVUE (HBN 1.5), the DPU will disregard the setting
and use the old way.

Also switch the api-tests integration test to using NVUE and
forge-dhcp-server, and make the mock `crictl` more realistic.