fix: use libredfish to power cycle Viking hosts. Only use the ipmi tool to reboot Vikings.
fixes: [https://nvbugs/4621846](https://nvbugs/4621846)

Using libredfish to reboot Viking servers also resets their DPU. So, we have to use the IPMI tool to reboot Viking servers.

However, we should use libredfish to reboot Viking servers.

Previously, we were using the IPMI tool to reboot Viking servers even though the caller was requested a change to the server's power status. Now, we use the IPMI tool for reboots and libredfish to toggle the power on Viking hosts.