chore(ci): only push the forge-admin-cli image when MR is merged
Currently every time a build pipeline runs it pushes a fresh
forge-admin-cli image to nvcr & urm. This means that the latest image
does not contain `jq` even though nvmetal/carbide!2091 was merged. This
would lead to failures of the scheduled test pipeline until all branches
in open MRs would contain the change. The goal of this commit is to
make the behaviour the same as for the other carbide containers, they
only get pushed when the MR is merged (or the special complete pipeline
tag appears in the commit message).

The two build jobs now optionally push.
The merge-manifests and copy to urm jobs are now only added to the
pipeline if the two build jobs pushed.