fix: Make --format=json work for single-host managed-host show
This makes JSON (and YAML) output work with `managed-host show <id>`, which used
to only work if showing all hosts.  Makes the `show_managed_hosts` fn
responsible for calling `show_managed_host_details_view` depending on whether
the user requested a single host or not.

It would be easier to just throw all found hosts at the show_managed_hosts fn,
but the goal is for it to render a single JSON object (not an array) if the user
wants one managed-host, hence the `single_host_detail_view` boolean.