feat: Support hosting multiple BMC mocks in a single address in machine-a-tron
This is currently incomplete (see TODO line in dev/machine-a-tron/src/bmc_mock_wrapper.rs), but it shows the plan I have in mind, which is to wrap bmc-mock's tar router with middleware which can intercept the request and rewrite the response. In the meantime, machine-a-tron still works, but you'll need to configure `override_target_{ip,port}` in carbide's site config (see comments in mat.toml) for it to work.

It also simplifies a bunch of stuff by fully embracing the "single mock host" technique for running bmc-mock, so:

- No more calling `ip addr add` since we don't need to listen on multiple IPs
- No more incrementing BMC ports from a base port
- HostMachine/DpuMachine aren't hosting bmc mocks individually any more
- No more redfish templates in machine-a-tron (deletes lots of JSON) since we're embracing the tar technique