fix: Get machine-a-tron working again
This (maybe temporarily?) abandons the approach for running a single
instance of BMC-mock for all mocked BMCs, because the carbide API no
longer respects the BMC target IP override option, and no longer sets
the correct `x-really-to-mac` headers needed to make that work. It's not
clear to me that we want to keep pursuing that option, as running
multiple BMC-mocks for multiple machines is not really a problem in
practice.

With this change we're now responding to site-explorer requests with
mock data that actually gives the correct MAC addresses of the mocked
machine/DPUs.

It also fixes several issues that were preventing machine-a-tron
from working:

- Send a `MachineValidationCompletedRequest` if the API state shows
  `MachineValidating`, simulating what forge-agent does, otherwise the
  state machine gets stuck

- Send a `MachineRebootCompletedRequest` after the mock host reboots,
  simulating what forge-agent does, otherwise the state machine gets
  stuck

- Respond to bios password set requests by sending a Location header of
  a valid-looking job ID, which is needed by the state machine when it
  reconfigures the BIOS password on Dell hosts

With this I have been able see machine-a-tron get hosts to the Ready
state once again.