chore(state_controller): Update ordering of redfish commands on DPU
The state DPUDiscovering/Configuring runs some Redfish commands to alter
the behavior of the DPU.  However, at this stage we can't guarantee that
any UEFI BIOS has posted or any OS has booted on the ARM side.

This reduces the things we do to the BIOS to the minimum set that
doesn't require any help from UEFI, and moves the general `forge_setup`
into DPUDiscovering/WaitingForPlatformConfiguration.

The minumum things is to get to:

- Disable Secure boot
- Configure UEFI HTTP Boot (we don't have have any idea of the interface
  names yet, so this will cause it to just go through the stages)