chore(api): configure measured boot generic db calls to return `DatabaseError`
Working through some TODOs re: measured boot cleanup. When I wrote all of
the code, I had a lot of `eyre::eyre!` throughout, and I planned on going back
through afterwords to figure out the right `Error` to drop in place. When I
finally went to go and do it, I was like, "oh geez, I've got a lot to clean
up". :laughing:

Starting by updating the generic DB calls in `common.rs` to return
`DatabaseError`, and then following the callers to update them as
well. There will be more!

One thing, btw, the `query` field for `DatabaseError` takes a `&'static str`,
which makes it hard to pass formatted query strings. BUT, it looks like the
`query` field generally isn't set as the query itself -- it looks like most
people just give it some sort of name/identifier. If it **should** be the
source query, then maybe we want to make it a `String` (similar to `GenericError`).