fix: Fix machine-a-tron when using nonstandard BMC ports
This fixes the production RedfishClientPool impl to respect the
override_target_port config for all redfish calls, not just
site-explorer ones.

Machine-a-tron was relying on machines sending `update_bmc_info` to the
API server in order for various redfish calls to use the correct port,
but this doesn't work for host machines, which don't run discovery and
thus don't send update_bmc_info.

With this change we don't need to rely on update_bmc_info to tell us
that the mock machines are running on a nonstandard port, instead the
whole API server will use the configured override port for all redfish
calls (this was already the case for site-explorer, but it wasn't
affecting things like BIOS password updates and triggering reboots.)