fix: otel change "log" attribute of frr.log from "frr" to "hbn"
This MR is just an attribute name change. Today we export
```
/var/log/doca/hbn/frr/frr.log
```
When I add other logs such as
```
/var/log/doca/hbn/nl2docad.log
/var/log/doca/hbn/syslog
/var/log/doca/hbn/supervisor/supervisord.log
```
I want them all to be identified as "hbn" logs. So I'm just changing the attribute from `log="frr"` to `log="hbn"`.

When I add more hbn logs, I expect I'll remove some of the parsed attributes from the frr.log, but I'm not planning to do that this week.