feat: Detect primary interface for Dell.
The logic to find primary DPU is designed based on booting interface. It does not matter which DPU is planned as primary, the one which is planned for booting host, is termed as primary. We observed that few Dell in demo2 are configured to boot via Slot.5-1 and others Slot.2-1. When we use the fix pci address based logic, host didn't come up.

After discussing, we came up with the idea to find the configured interface for booting, get the Mac and match it with host interface Mac while creating machine_interface. The one matches will be used as primary interface.

The benefit is that carbide is not bound to some restriction. Carbide can take care of any DPU based on booting interface. If DPU0 becomes faulty, DPU1 can become Primary based on booting interface (of course after force-delete).

Later on carbide can fix the host's booting interface and use only DPU connected to that interface as primary, e.g. for Dell NIC.Slot.5-1. The current approach can be used to restrict booting interface as per configuration. This configuration can be as per site, or provided in `expected_machine` table.